Move UI-only things to Fluent project
