+++ WalletWasabi/WabiSabi/Client/InternalDestinationProvider.cs
@@ -18,14 +18,14 @@ public class InternalDestinationProvider : IDestinationProvider
 	{
 		// Get all locked internal keys we have and assert we have enough.
 		KeyManager.AssertLockedInternalKeysIndexedAndPersist(count, preferTaproot);
-		var preferScriptPubKeyType = preferTaproot 
+		var preferedScriptPubKeyType = preferTaproot 
 			? ScriptPubKeyType.TaprootBIP86 
 			: ScriptPubKeyType.Segwit;
 
 		bool IsAvailable(HdPubKey hdPubKey) =>
 			hdPubKey.IsInternal &&
 			hdPubKey.KeyState == KeyState.Locked &&
-			hdPubKey.FullKeyPath.GetScriptTypeFromKeyPath() == preferScriptPubKeyType;
+			hdPubKey.FullKeyPath.GetScriptTypeFromKeyPath() == preferedScriptPubKeyType;
 		
 		return KeyManager
 			.GetKeys(IsAvailable)
